+++ conwaymd/employables.py
@@ -1708,6 +1708,7 @@ class ExplicitLinkReplacement(
 
 
 class SpecifiedLinkReplacement(
+    ReplacementWithAllowedFlags,
     ReplacementWithAttributeSpecifications,
     ReplacementWithProhibitedContent,
     Replacement,
@@ -1719,6 +1720,7 @@ class SpecifiedLinkReplacement(
     ````
     SpecifiedLinkReplacement: #«id»
     - queue_position: (def) NONE | ROOT | BEFORE #«id» | AFTER #«id»
+    - allowed_flags: (def) NONE | «letter»=«FLAG_NAME» [...]
     - attribute_specifications: (def) NONE | EMPTY | «string»
     - prohibited_content: (def) NONE | BLOCKS | ANCHORED_BLOCKS
     ````
@@ -1735,6 +1737,7 @@ class SpecifiedLinkReplacement(
     def attribute_names(self) -> tuple[str, ...]:
         return (
             'queue_position',
+            'allowed_flags',
             'attribute_specifications',
             'prohibited_content',
         )
@@ -1743,14 +1746,21 @@ class SpecifiedLinkReplacement(
         pass
 
     def _set_apply_method_variables(self):
+        self._has_flags = len(self._flag_name_from_letter) > 0
         self._regex_pattern_compiled = re.compile(
             pattern=SpecifiedLinkReplacement.build_regex_pattern(
+                self._flag_name_from_letter,
+                self._has_flags,
                 self._attribute_specifications,
                 self._prohibited_content_regex,
             ),
             flags=re.ASCII | re.VERBOSE,
         )
-        self._substitute_function = SpecifiedLinkReplacement.build_substitute_function(self._attribute_specifications)
+        self._substitute_function = SpecifiedLinkReplacement.build_substitute_function(
+            self._flag_name_from_letter,
+            self._has_flags,
+            self._attribute_specifications,
+        )
 
     def _apply(self, string: str) -> str:
         return re.sub(
@@ -1760,7 +1770,9 @@ class SpecifiedLinkReplacement(
         )
 
     @staticmethod
-    def build_regex_pattern(attribute_specifications: Optional[str], prohibited_content_regex: Optional[str]) -> str:
+    def build_regex_pattern(flag_name_from_letter: dict[str, str], has_flags: bool,
+                            attribute_specifications: Optional[str], prohibited_content_regex: Optional[str]) -> str:
+        flags_regex = build_flags_regex(flag_name_from_letter, has_flags)
         link_text_regex = build_content_regex(prohibited_content_regex,
                                               permitted_content_regex=r'[^\]]', capture_group_name='link_text')
         bracketed_link_text_regex = fr'\[ [\s]* {link_text_regex} [\s]* \]'
@@ -1775,6 +1787,7 @@ class SpecifiedLinkReplacement(
         closing_parenthesis_regex = r'\)'
 
         return ''.join([
+            flags_regex,
             bracketed_link_text_regex,
             attribute_specifications_regex,
             opening_parenthesis_regex,
@@ -1785,8 +1798,12 @@ class SpecifiedLinkReplacement(
         ])
 
     @staticmethod
-    def build_substitute_function(attribute_specifications: Optional[str]) -> Callable[[re.Match], str]:
+    def build_substitute_function(flag_name_from_letter: dict[str, str], has_flags: bool,
+                                  attribute_specifications: Optional[str]) -> Callable[[re.Match], str]:
         def substitute_function(match: re.Match) -> str:
+            enabled_flag_names = ReplacementWithAllowedFlags.get_enabled_flag_names(match,
+                                                                                    flag_name_from_letter, has_flags)
+
             content = match.group('link_text')
 
             angle_bracketed_uri = match.group('angle_bracketed_uri')